Build static library with fvisibility=hidden.
There were several reports of Anaconda's MKL interfering with extensions that links against fftw in unexpected ways.

https://github.com/pyFFTW/pyFFTW/issues/40
https://github.com/openjournals/joss-reviews/issues/1340

MKL runtime overrides `fftw*` dynamic symbols exported by the extension module. If the symbols in the .a archive are not marked as hidden, then they will be exported as dynamic symbols in the generated so file. Therefore with the current visibility setting even if a downstream extension links against the static version of the fftw library, MKL can still intervene and break the extensions.

With this PR, if a downstream extension links against the static fftw libraries, then MKL will not break them.